Add special ! and - markers
The ! marker is used to indicate that searches should not continue
upward in the directory tree (however, both .gocfg and markers will be
checked for the directory that was stopped on, regardless of which
stopped it).

The - marker says to drop GOPATH entries that are already in the
environment.

Also adds a -verbose option for debugging gopath. Not really something
you'd turn on normally.